♻️ refactor(database): move transfer enums to @lobechat/types and drop schema-level enum typing
