+++ packages/database/src/schemas/resourceTransferRequest.ts
@@ -1,3 +1,8 @@
+import type {
+  ResourceTransferRequestOptions,
+  ResourceTransferRequestStatus,
+  TransferResourceType,
+} from '@lobechat/types';
 import { sql } from 'drizzle-orm';
 import { index, jsonb, pgTable, text, uniqueIndex, uuid } from 'drizzle-orm/pg-core';
 
@@ -6,61 +11,21 @@ import { users } from './user';
 import { workspaces } from './workspace';
 
 /**
- * Entity kinds that support member-to-member ownership transfer. Polymorphic on
- * purpose (mirroring `resource_permissions`): onboarding a new entity only
- * requires a new literal here plus an accept-executor, not a new table.
- * v1 wires up `agent` only; the remaining literals are reserved for the
- * follow-up integrations.
- */
-export const TRANSFER_RESOURCE_TYPES = [
-  'agent',
-  'agentGroup',
-  'document',
-  'file',
-  'knowledgeBase',
-] as const;
-export type TransferResourceType = (typeof TRANSFER_RESOURCE_TYPES)[number];
-
-/**
- * Lifecycle of a transfer request. `pending` is the only live state; every
- * other state is terminal:
- * - `accepted`  — recipient confirmed, ownership has been handed over
- * - `declined`  — recipient refused
- * - `cancelled` — initiator withdrew, or the resource left the workspace /
- *                 was deleted before the recipient answered
- * - `expired`   — nobody acted before `expiresAt` (stamped lazily on read)
- */
-export const RESOURCE_TRANSFER_REQUEST_STATUSES = [
-  'pending',
-  'accepted',
-  'declined',
-  'cancelled',
-  'expired',
-] as const;
-export type ResourceTransferRequestStatus = (typeof RESOURCE_TRANSFER_REQUEST_STATUSES)[number];
-
-export interface ResourceTransferRequestOptions {
-  /**
-   * Hand the initiator's own topics/messages of this resource to the recipient
-   * on accept. Only the resource creator may set it (a primary owner
-   * reassigning someone else's resource cannot give away conversations that
-   * are not theirs).
-   */
-  migrateSessions?: boolean;
-}
-
-/**
  * Member-to-member ownership handover requests inside one workspace, gated on
  * the recipient's confirmation. One row per initiated transfer; the resource's
  * ownership column (`agents.userId`, …) only changes when the recipient
  * accepts.
+ *
+ * `resourceType` / `status` are plain text columns typed via `@lobechat/types`
+ * unions — no DB-level enum, so onboarding a new resource kind or state never
+ * needs a migration.
  */
 export const resourceTransferRequests = pgTable(
   'resource_transfer_requests',
   {
     id: uuid('id').defaultRandom().primaryKey().notNull(),
 
-    resourceType: text('resource_type', { enum: TRANSFER_RESOURCE_TYPES }).notNull(),
+    resourceType: text('resource_type').$type<TransferResourceType>().notNull(),
     resourceId: text('resource_id').notNull(),
 
     workspaceId: text('workspace_id')
@@ -86,9 +51,7 @@ export const resourceTransferRequests = pgTable(
       onDelete: 'set null',
     }),
 
-    status: text('status', { enum: RESOURCE_TRANSFER_REQUEST_STATUSES })
-      .notNull()
-      .default('pending'),
+    status: text('status').$type<ResourceTransferRequestStatus>().notNull().default('pending'),
 
     options: jsonb('options').$type<ResourceTransferRequestOptions>(),
