+++ packages/types/src/index.ts
@@ -36,6 +36,7 @@ export * from './plugins';
 export * from './project';
 export * from './rag';
 export * from './redis';
+export * from './resourceTransfer';
 export * from './search';
 export * from './serverConfig';
 export * from './service';
